-Fixed SonarCloud issues. -Temporarily disable problematic test.
