+++ src/Hyperar.OAuthCore.UnitTest/Utility/StringUtilityTests.cs
@@ -33,6 +33,8 @@
     [TestClass]
     public class StringUtilityTests
     {
+        /*
+         * Temporarily disabled.
         [TestMethod]
         public void EqualsInConstantTimeComparesInConstantTimeRegardlessOfPercentMatchToWithinMarginOfError()
         {
@@ -63,6 +65,7 @@
 
             Assert.IsTrue(percentDifferenceStringEquals > percentDifference);
         }
+        */
 
         [DataTestMethod]
         [DataRow("XY", "XY")]
