+++ src/Hyperar.OAuthCore/Framework/Signing/RsaSha1SignatureImplementation.cs
@@ -41,6 +41,8 @@ namespace Hyperar.OAuthCore.Framework.Signing
 
         public bool ValidateSignature(IOAuthContext? authContext, SigningContext? signingContext)
         {
+            ArgumentNullException.ThrowIfNull(authContext);
+            ArgumentNullException.ThrowIfNull(signingContext);
             ArgumentException.ThrowIfNullOrWhiteSpace(authContext?.Signature);
             ArgumentException.ThrowIfNullOrWhiteSpace(signingContext?.SignatureBase);
